Caesar Cipher is functional. Webpage done.
The webpage is completely functional. Caesar cipher is applied to an individual letter in the shiftLetter function. shiftLetter function ignores whitespace, special symbol and numbers. shiftLetter is applied to each letter in the first textbox and the resulting cipher shown in the second textbox using the EncryptText function.

The page now has an input for numbers. Allows for positive or negative
whole numbers. Changing the input changes the prompt in the middle of
the page and updates any cipher in the bottom textarea.

Updated the css to keep all textarea and input in the middle. Format in
the order of id, class, then elements.